+++ lib/chewy/type/witchcraft.rb
@@ -60,7 +60,7 @@ module Chewy
         def alicorn
           @alicorn ||= class_eval <<-RUBY
             -> (locals, object0, crutches) do
-              #{composed_values(@type.root_object, 0)}
+              #{composed_values(@type.root, 0)}
             end
           RUBY
         end