+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changes
 
+  * Do not underscore suggested index name on `Chewy::Index.index_name` call.
   * It is possible now to call `root` method several times inside a single type definition, the options will be merged. Also, the block isn't required anymore.
 
 # Version 0.10.1